requirements.txt
pure-python-adb
Pillow
pytesseract
progress